v0.2.187: 種花模式「每點圈數」支援 0.5 步進(0.5=半圈,可設 0.5/1.5/2.5…)

+++ backend/core/simulation_engine.py
@@ -338,7 +338,7 @@ class SimulationEngine:
         mode: MovementMode,
         radius_m: float = 30.0,
         segments: int = 8,
-        circles: int = 1,
+        circles: float = 1.0,
         rounds: int = 1,
         pre_wait: float = 3.0,
         post_wait: float = 3.0,
